fix: quota_left field type to int8
(cherry picked from commit 1b752f7e8d65384d7b791a47f00f28fe586544ce)

+++ server/cmd/plan/service/pack.go
@@ -44,6 +44,6 @@ func packPlanView(p *model.PlanView) *plan.Plan {
 		Place:          &p.Place,
 		Message:        p.Message,
 		Quota:          &p.Quota,
-		RemainingQuota: lo.ToPtr(int8(p.QuotaLeft)),
+		RemainingQuota: lo.ToPtr(p.QuotaLeft),
 	}
 }
